+++ changelog.md
@@ -8,6 +8,8 @@
 
 **Feature:** Add output "lb_arn"
 
+**Change:** All policies are now rendered with `aws_iam_policy_document` rather than as json blobs
+
 # 4.6
 
 **Bugfix:** Change quote style for ssh_populate scripts to avoid early interpolation (Thanks KevinGreen)